fixing broken keyword argument

+++ research/scripts/multi_sim_script.py
@@ -61,21 +61,6 @@ def make_true(given_key):
         true_scale = test_info['params']['true_scale']
         true_nz = gamma(true_loc, true_scale**2, bounds=(min(test_info['bin_ends']), max(test_info['bin_ends'])))#sps.erlang(true_shape, true_loc, true_scale)
         true_dict = {'amps': [1.], 'means': [true_loc], 'sigmas': [true_scale]}
-            # true_amps = np.array([0.150,0.822,1.837,2.815,3.909,
-            #                       5.116,6.065,6.477,6.834,7.304,
-            #                       7.068,6.771,6.587,6.089,5.165,
-            #                       4.729,4.228,3.664,3.078,2.604,
-            #                       2.130,1.683,1.348,0.977,0.703,
-            #                       0.521,0.339,0.283,0.187,0.141,
-            #                       0.104,0.081,0.055,0.043,0.034])
-            # true_grid = np.linspace(test_info['bin_ends'][0], test_info['bin_ends'][-1], len(true_amps) + 1)
-            # true_grid_mids = (true_grid[1:] + true_grid[:-1]) / 2.
-            # f = spi.interp1d(true_grid_mids, true_amps)
-            # bin_mids = (test_info['bin_ends'][1:] + test_info['bin_ends'][:-1]) / 2.
-            # bin_difs = test_info['bin_ends'][1:] - test_info['bin_ends'][:-1]
-            # true_means = bin_mids
-            # true_amps = f(bin_mids)
-            # true_sigmas = bin_difs
     else:
         bin_range = max(test_info['bin_ends']) - min(test_info['bin_ends'])
         true_amps = np.array([0.20, 0.35, 0.55])
@@ -124,7 +109,7 @@ def make_interim_prior(given_key):
             int_funcs.append(chippr.gauss(int_means[c], int_sigmas[c]**2))
         interim_prior = chippr.gmix(int_amps, int_funcs,
             limits=(min(test_info['bin_ends']), max(test_info['bin_ends'])))
-    elif test_info['params']['interim_prior'] == 'training':
+    elif int_pr_type == 'training':
         int_amps = np.array([0.150,0.822,1.837,2.815,3.909,
                               5.116,6.065,6.477,6.834,7.304,
                               7.068,6.771,6.587,6.089,5.165,
@@ -132,12 +117,20 @@ def make_interim_prior(given_key):
                               2.130,1.683,1.348,0.977,0.703,
                               0.521,0.339,0.283,0.187,0.141,
                               0.104,0.081,0.055,0.043,0.034])
-        int_grid = np.linspace(test_info['bin_ends'][0], test_info['bin_ends'][-1], len(int_amps) + 1)
-        int_grid_mids = (int_grid[1:] + int_grid[:-1]) / 2.
-        f = spi.interp1d(int_grid_mids, int_amps)
-        bin_mids = (test_info['bin_ends'][1:] + test_info['bin_ends'][:-1]) / 2.
-        bin_difs = test_info['bin_ends'][1:] - test_info['bin_ends'][:-1]
-        int_amps = f(bin_mids)
+        int_grid = np.linspace(0., 1.1, len(int_amps) + 1)
+        int_mids = (int_grid[1:] + int_grid[:-1]) / 2.
+        int_amps = spi.griddata(int_mids, int_amps, bin_mids, method='linear', fill_value=np.min(int_amps), rescale=False)
+        # # int_grid = np.concatenate((int_grid, np.array([test_info['bin_ends'][-1]])))
+        # # print(int_grid)
+        # int_amps.append(int_amps[-1])
+        # int_amps = np.array(int_amps)
+        # # this basically hardcodes in that grids start at 0.!
+        #
+        # int_grid_mids = (int_grid[1:] + int_grid[:-1]) / 2.
+        #
+        # int_grid_mids = np.concatenate((int_grid_mids, np.array([test_info['bin_ends'][-1]])))
+        # f = spi.interp1d(int_grid_mids, int_amps)
+        # int_amps = f(bin_mids)
         int_means = bin_mids
         int_sigmas = bin_difs
         n_mix_comps = len(int_amps)
@@ -167,9 +160,12 @@ def make_catalog(given_key):
 
     test_dir = os.path.join(result_dir, test_name)
     test_info['dir'] = test_dir
-    if os.path.exists(test_dir):
-        shutil.rmtree(test_dir)
-    os.makedirs(test_dir)
+    if test_dir != '':
+        if os.path.exists(test_dir):
+            shutil.rmtree(test_dir)
+        os.makedirs(test_dir)
+    else:
+        print('Check the config file for newlines!')
 
     param_file_name = test_name + '.txt'
     params = chippr.utils.ingest(param_file_name)
@@ -195,7 +191,7 @@ def make_catalog(given_key):
     interim_prior = make_interim_prior(given_key)
 
     posteriors = chippr.catalog(param_file_name, loc=test_dir, prepend=test_name)
-    output = posteriors.create(true_nz, interim_prior, N=test_info['params']['n_gals'])
+    output = posteriors.create(true_nz, interim_prior, N=test_info['params']['n_gals'], vb=True)
     # data = np.exp(output['log_interim_posteriors'])
     posteriors.write()
     data_dir = posteriors.data_dir
